requirements.txt
torch>=1.12.0
torchvision>=0.13.0
diffusers>=0.14.0
transformers>=4.25.1
Pillow>=9.3.0
tqdm>=4.64.1
peft>=0.2.0
accelerate>=0.16.0
safetensors>=0.3.0